+++ tests/Persisters/Serializers/RBXETest.php
@@ -5,16 +5,19 @@ namespace Rubix\ML\Tests\Persisters\Serializers;
 use Rubix\ML\Encoding;
 use Rubix\ML\Persistable;
 use Rubix\ML\Classifiers\DummyClassifier;
-use Rubix\ML\Persisters\Serializers\Bzip2;
+use Rubix\ML\Persisters\Serializers\RBXE;
 use Rubix\ML\Persisters\Serializers\Serializer;
 use PHPUnit\Framework\TestCase;
+use Rubix\ML\Exceptions\RuntimeException;
+use stdClass;
+
+use function serialize;
 
 /**
  * @group Serializers
- * @requires extension bz2
- * @covers \Rubix\ML\Persisters\Serializers\Bzip2
+ * @covers \Rubix\ML\Persisters\Serializers\RBXE
  */
-class Bzip2Test extends TestCase
+class RBXETest extends TestCase
 {
     /**
      * @var \Rubix\ML\Persistable
@@ -22,7 +25,7 @@ class Bzip2Test extends TestCase
     protected $persistable;
 
     /**
-     * @var \Rubix\ML\Persisters\Serializers\Bzip2
+     * @var \Rubix\ML\Persisters\Serializers\RBXE
      */
     protected $serializer;
 
@@ -31,7 +34,7 @@ class Bzip2Test extends TestCase
      */
     protected function setUp() : void
     {
-        $this->serializer = new Bzip2(4, 0);
+        $this->serializer = new RBXE('secret');
 
         $this->persistable = new DummyClassifier();
     }
@@ -41,7 +44,7 @@ class Bzip2Test extends TestCase
      */
     public function build() : void
     {
-        $this->assertInstanceOf(Bzip2::class, $this->serializer);
+        $this->assertInstanceOf(RBXE::class, $this->serializer);
         $this->assertInstanceOf(Serializer::class, $this->serializer);
     }
 
@@ -58,5 +61,32 @@ class Bzip2Test extends TestCase
 
         $this->assertInstanceOf(DummyClassifier::class, $persistable);
         $this->assertInstanceOf(Persistable::class, $persistable);
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function deserializeInvalidData() : array
+    {
+        return [
+            [3],
+            [new stdClass()],
+        ];
+    }
+
+    /**
+     * @test
+     *
+     * @param mixed $obj
+     *
+     * @dataProvider deserializeInvalidData
+     */
+    public function deserializeBadData($obj) : void
+    {
+        $data = new Encoding(serialize($obj));
+
+        $this->expectException(RuntimeException::class);
+
+        $this->serializer->unserialize($data);
     }
 }